requirements.txt
aleph_alpha_client
python-dotenv
langchain
